runtime/ops/pyproject.toml
[project]
name = "ops"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "datasketch>=1.8.0",
    "email-validator>=2.3.0",
    "emoji>=2.15.0",
    "jieba>=0.42.1",
    "loguru>=0.7.3",
    "numpy>=2.2.0,<=2.2.6",
    "opencv-contrib-python-headless>=4.12.0.88",
    "opencv-python-headless>=4.12.0.88",
    "openslide-python>=1.4.3",
    "paddleocr>=3.3.2",
    "pandas>=2.2.0,<=2.2.3",
    "pycryptodome>=3.23.0",
    "pymysql>=1.1.2",
    "python-docx>=1.2.0",
    "pytz>=2025.2",
    "six>=1.17.0",
    "sqlalchemy>=2.0.44",
    "xmltodict>=1.0.2",
    "zhconv>=1.4.3",
]